% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPA.R
\name{SPA}
\alias{SPA}
\title{Successive projection algorithm for separable NMF}
\usage{
SPA(X, r, options = list())
}
\arguments{
\item{X}{an m-by-n matrix.
Ideally admitting a near-separable factorization, that is, X = WH + N, where

conv([W, 0]) has r vertices,

H = [I,H']P where I is the identity matrix, H'>= 0 and its
columns sum to at most one, P is a permutation matrix, and

N is sufficiently small.}

\item{r}{number of columns to be extracted.}

\item{options}{list of string
if options contain

"normalize" := 1, will scale the columns of X so that they sum to one,
hence matrix H will satisfy the assumption above for any
nonnegative separable matrix X.

"normalize" := 0, is the default value for which no scaling is
performed. For example, in hyperspectral imaging, this
assumption is already satisfied and normalization is not
necessary.

"precision" : stops the algorithm when
max_k ||R(:,k)||_2 <= relerror * max_k ||X(:,k)||_2
where R is the residual, that is, R = X-X(:,K)H.
default: 1e-6

"display"   : = 1, displays the iteration count (default)}
}
\value{
\code{K} index set of the extracted columns.
}
\description{
At each step of the algorithm, the column of R maximizing ||.||_2 is
extracted, and R is updated by projecting its columns onto the orthogonal
complement of the extracted column. The residual R is initializd with X.

See N. Gillis and S.A. Vavasis, Fast and Robust Recursive Algorithms
for Separable Nonnegative Matrix Factorization,  IEEE Trans. on Pattern
Analysis and Machine Intelligence 36 (4), pp. 698-714, 2014.
}
\examples{
SPA(matrix(1:16,4),2,c('normalize'))

}
